% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ce}
\alias{ce}
\title{cell embeddings for pbmc3k data}
\format{
A list
}
\source{
{created with pbmc3k data}
}
\usage{
ce
}
\description{
cell embeddings for pbmc3k data
}
\examples{
data(ce)       #Lazy loading. Data becomes visible as soon as called
}
\keyword{datasets}
