% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{collect.random.set}
\alias{collect.random.set}
\title{6. find gene sets with spatial relevance}
\usage{
\method{collect}{random.set}(n.genes, n.times, gene.use, el)
}
\arguments{
\item{n.genes}{number of genes}

\item{n.times}{how many simulated gene sets to generate}

\item{gene.use}{background genes; usually all genes in the dataset}

\item{el}{this is used for filtering genes to avoid the introduce of genes that did not pass
the MCA processing}
}
\value{
list of gene lists
}
\description{
This function is called by 'compute.random.pal.list' for collecting random gene sets
}
