% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{compute.cell.label.df}
\alias{compute.cell.label.df}
\title{similar to compute.cell.label; used when working with multiple gene sets}
\usage{
\method{compute}{cell.label.df}(cell_df)
}
\arguments{
\item{cell_df:}{output of 'run.rwr.list'}
}
\description{
similar to compute.cell.label; used when working with multiple gene sets
}
\examples{
cell.prob.df <- run.rwr.list(el = el, gene_set_list = gene.set.list, cells = cells)
cell.label.df <- compute.cell.label.df(cell.prob.df)
}
