% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{compute.cell.label}
\alias{compute.cell.label}
\title{4.2. binarize the label propagation probability in the cell population;
result in a binarized vector of cells with 'nagative' and 'positive' labels;
'positive' means that the cells are relevant to the gene set}
\usage{
\method{compute}{cell.label}(cell_vec)
}
\arguments{
\item{cell_vec}{output of 'run.rwr'}
}
\value{
cell label of 'negative' or 'positive' for a given pathway
}
\description{
4.2. binarize the label propagation probability in the cell population;
result in a binarized vector of cells with 'nagative' and 'positive' labels;
'positive' means that the cells are relevant to the gene set
}
\examples{
\donttest{
cells <- colnames(pbmc.mtx)
el <- gsdensity::compute.nn.edges(coembed = ce, nn.use = 300)
cv <- gsdensity::run.rwr(el = el,
                         gene_set = gene.set.list[["GOBP_B_CELL_ACTIVATION"]],
                         cells = cells)
cl <- compute.cell.label(cv)
}

}
