% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{compute.db}
\alias{compute.db}
\title{this function is called by 'compute.kld' to aggregate the density contribution
of each gene to each grid point, and then normalize the densities of grid points
to 1.}
\usage{
\method{compute}{db}(density.df)
}
\description{
this function is called by 'compute.kld' to aggregate the density contribution
of each gene to each grid point, and then normalize the densities of grid points
to 1.
}
