% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{compute.grid.coords}
\alias{compute.grid.coords}
\title{2. compute density of gene sets of interest
2.1 compute grid point coordinates}
\usage{
\method{compute}{grid.coords}(coembed, genes.use, n.grids = 100)
}
\arguments{
\item{coembed}{the result from compute.mca}

\item{genes.use}{which genes to use; no default;
can use genes based on the gene set selection or use rownames(object)}

\item{n.grids}{number of grid points used for gene set density estimation;
larger number is more accurate and slower;
default is 100 (recommended to test 100 first)}
}
\value{
grid coordinates
}
\description{
2. compute density of gene sets of interest
2.1 compute grid point coordinates
}
\examples{

compute.grid.coords(coembed = ce, genes.use = intersect(rownames(ce), rownames(pbmc.mtx)))

}
