% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{compute.jsd}
\alias{compute.jsd}
\title{5. compute the specificity of gene set when cell partition information is available;
the information could be clustering, sample origins, or other conditions
inspired by https://github.com/FloWuenne/scFunctions/blob/0d9ea609fa72210a151f7270e61bdee008e8fc88/R/calculate_rrs.R}
\usage{
\method{compute}{jsd}(x, y)
}
\description{
this function is called by compute.spec.single to calculate the similarity
between two vectors
}
