% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{compute.nn.edges}
\alias{compute.nn.edges}
\title{3. compute nearest neighbor graph for genes and cells
This graph will be used for fetching the most relevant cells of a gene set}
\usage{
\method{compute}{nn.edges}(coembed, nn.use = 300)
}
\arguments{
\item{coembed}{the result from compute.mca}

\item{nn.use}{the number of nearest neighbors for building the graph; default 300}
}
\value{
nearest neighbor graph (edges)
}
\description{
3. compute nearest neighbor graph for genes and cells
This graph will be used for fetching the most relevant cells of a gene set
}
\examples{
\donttest{
compute.nn.edges(coembed = ce)
}

}
