% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{compute.random.pal.list}
\alias{compute.random.pal.list}
\title{This function generates pathway activity calculations for random gene sets;
to save computational resource, we would not simulate random gene sets for each real gene set (for size matching);
instead, we will group the gene sets with similar sizes (by 10, e.g., 1-10, 11-20, 21-30,...) and use a random gene
set (with size 6, 16, 26, ...) for all the real gene sets within that range.}
\usage{
\method{compute}{random.pal.list}(gene.set.list, n.times, gene.use, el, cells, parallel = TRUE)
}
\arguments{
\item{gene.set.list}{real gene list of interest}

\item{n.times}{how many simulated gene sets to generate}

\item{gene.use}{background genes; usually all genes in the dataset}

\item{el}{this is used for filtering genes to avoid the introduce of genes that did not pass
the MCA processing}

\item{cells}{cells; this will be used for run.rwr}

\item{parallel}{parallel parameter for dRWR; default is true}
}
\value{
list of two objects; first, a list of pathway acticity dataframes (for each random gene set);
second, a matching file between each real gene set and each random gene set pathway acticities
}
\description{
This function generates pathway activity calculations for random gene sets;
to save computational resource, we would not simulate random gene sets for each real gene set (for size matching);
instead, we will group the gene sets with similar sizes (by 10, e.g., 1-10, 11-20, 21-30,...) and use a random gene
set (with size 6, 16, 26, ...) for all the real gene sets within that range.
}
