% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{compute.spatial.kld.df}
\alias{compute.spatial.kld.df}
\title{This function is a parallel mode of compute.spatial.kld for testing multiple gene sets
to calculate how likely the cells relevant to a gene set is
randomly distributed spatially. Done by comparison between a real gene list and multiple
random gene sets (size matched)}
\usage{
\method{compute}{spatial.kld.df}(spatial.coords, weight_df, r.df, n = 10)
}
\arguments{
\item{spatial.coords}{a data frame with each row as a cell and each column
as a spatial coordinate (usually 2: x and y)}

\item{r.df}{output of compute.random.pal.list}

\item{n}{split the spatial map for local density estimation;
n is the number of split for each dimension; for n = 10, the spatial map is
split to n * n = 100 grids for the density estimation}

\item{weight_vec}{output of run.rwr}
}
\value{
spatial kl-divergence and statistics
}
\description{
This function is a parallel mode of compute.spatial.kld for testing multiple gene sets
to calculate how likely the cells relevant to a gene set is
randomly distributed spatially. Done by comparison between a real gene list and multiple
random gene sets (size matched)
}
\examples{
\donttest{
compute.spatial.kld.df(spatial.coords = coords.df,r.df = random.activities,
weight_df = weight_df, n = 10)
}
}
