% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{compute.spec}
\alias{compute.spec}
\title{This is to calculate the similarity between:
1. the label propagation probability of cells for gene sets and
2. the identify of cells in partitions}
\usage{
\method{compute}{spec}(cell_df, metadata, cell_group)
}
\arguments{
\item{metadata:}{a data frame with cell information (each row is a cell;
usually object@meta.data)}

\item{cell_group:}{cell partition vector (usually a column name
in object@meta.data)}

\item{cell_df:}{the output of run.rwr.list}
}
\description{
This is to calculate the similarity between:
1. the label propagation probability of cells for gene sets and
2. the identify of cells in partitions
}
\examples{
jsd.df <- compute.spec(cell_df = cell.prob.df,
                       metadata = object@meta.data,
                       cell_group = "seurat_clusters")
}
