% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coords.df}
\alias{coords.df}
\title{mouse brain coords}
\format{
A df
}
\source{
{created with brain data}
}
\usage{
coords.df
}
\description{
mouse brain coords
}
\examples{
data(coords.df)       #Lazy loading. Data becomes visible as soon as called
}
\keyword{datasets}
