% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene.set.list}
\alias{gene.set.list}
\title{A gene set list containing multiple human GO gene sets}
\format{
A list
}
\source{
{created based on msigdbr human C5 (biological process) gene sets}
}
\usage{
gene.set.list
}
\description{
A gene set list containing multiple human GO gene sets
}
\examples{
data(gene.set.list)       #Lazy loading. Data becomes visible as soon as called
}
\keyword{datasets}
