% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.meta}
\alias{pbmc.meta}
\title{pbmc3k meta}
\format{
A df
}
\source{
{created with pbmc3k data}
}
\usage{
pbmc.meta
}
\description{
pbmc3k meta
}
\examples{
data(pbmc.meta)       #Lazy loading. Data becomes visible as soon as called
}
\keyword{datasets}
