% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.mtx}
\alias{pbmc.mtx}
\title{pbmc3k matrix}
\format{
A matrix
}
\source{
{created with pbmc3k data}
}
\usage{
pbmc.mtx
}
\description{
pbmc3k matrix
}
\examples{
data(pbmc.mtx)       #Lazy loading. Data becomes visible as soon as called
}
\keyword{datasets}
