% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{random.activities}
\alias{random.activities}
\title{random gene set activities- matching weight_df}
\format{
A list
}
\source{
{created with brain data}
}
\usage{
random.activities
}
\description{
random gene set activities- matching weight_df
}
\examples{
data(random.activities)       #Lazy loading. Data becomes visible as soon as called
}
\keyword{datasets}
