% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{run.rwr}
\alias{run.rwr}
\title{4.1 To calculate the label propagation probability for a gene set among cells;
result in a vector (length = number of cells) reflecting the probability
each cell is labeled during the propagation (relevance to the gene set)}
\usage{
\method{run}{rwr}(el, gene_set, cells)
}
\arguments{
\item{el:}{edge list; output of 'compute.nn.edges'}

\item{gene_set:}{a vector of genes of interest}

\item{cells:}{name of cells; usually the same as 'colnames(object)'}
}
\description{
4.1 To calculate the label propagation probability for a gene set among cells;
result in a vector (length = number of cells) reflecting the probability
each cell is labeled during the propagation (relevance to the gene set)
}
\examples{
cells <- colnames(pbmc3k)
el <- compute.nn.edges(coembed = ce, nn.use = 300)
gi <- "GOBP_TAXIS"
cv <- run.rwr(el = el, gene_set = gene.set.list[[gi]], cells = cells)
}
