% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{sample.kld}
\alias{sample.kld}
\title{this function is called by 'compute.kld' to calculate the kl-divergence between
sampled (background) gene set and the ref (all) gene set}
\usage{
sample.kld(density.df, ref, len.gene.set)
}
\description{
this function is called by 'compute.kld' to calculate the kl-divergence between
sampled (background) gene set and the ref (all) gene set
}
