% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{seed.mat.list}
\alias{seed.mat.list}
\title{this function is used when more than one 'seed sets' will be used (when there
are multiple gene sets of interest)}
\usage{
seed.mat.list(gene_set_list, graph.use)
}
\arguments{
\item{gene_set_list}{gene_set_list}

\item{graph.use}{graph.use}
}
\value{
returns seed matrix
}
\description{
this function is used when more than one 'seed sets' will be used (when there
are multiple gene sets of interest)
}
