% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{vectorized_pdist}
\alias{vectorized_pdist}
\title{from an excellent post: https://www.r-bloggers.com/2013/05/pairwise-distances-in-r/
enhanced the speed
this function is called by 'compute.kld' to quickly compute the distance between
genes to grid points}
\usage{
vectorized_pdist(A, B)
}
\description{
from an excellent post: https://www.r-bloggers.com/2013/05/pairwise-distances-in-r/
enhanced the speed
this function is called by 'compute.kld' to quickly compute the distance between
genes to grid points
}
