% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight_df}
\alias{weight_df}
\title{mouse brain gene set activities}
\format{
A df
}
\source{
{created with brain data}
}
\usage{
weight_df
}
\description{
mouse brain gene set activities
}
\examples{
data(weight_df)       #Lazy loading. Data becomes visible as soon as called
}
\keyword{datasets}
