% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{el_nn_search}
\alias{el_nn_search}
\title{this function is called by 'compute.nn.edges' to convert nearest neighbor
identity matrix to edge list}
\usage{
el_nn_search(nn2_out)
}
\description{
this function is called by 'compute.nn.edges' to convert nearest neighbor
identity matrix to edge list
}
