% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdensity_functions.R
\name{sample.spatial.kld}
\alias{sample.spatial.kld}
\title{this function is called by 'compute.spatial.kld' to calculate the kl-divergence between
cell-weighted with shuffled weight vector and the ref (all cells, unweighted)}
\usage{
sample.spatial.kld(weight_vec, spatial.coords, n, ref)
}
\description{
this function is called by 'compute.spatial.kld' to calculate the kl-divergence between
cell-weighted with shuffled weight vector and the ref (all cells, unweighted)
}
